Rename git to lib
