--- src/cmd/git/mod.ts
@@ -1,6 +1,0 @@
-export { cloneCmd } from "./clone.ts";
-export { pullCmd } from "./pull.ts";
-export { statusCmd } from "./status.ts";
-export { branchCmd } from "./branch.ts";
-export { checkoutCmd } from "./checkout.ts";
-export { pushCmd } from "./push.ts";
