+++ src/cmd/lib/clone.ts
