+++ src/cmd/lib/mod.ts
@@ -1,0 +1,9 @@
+export { branchCmd } from "./branch.ts";
+export { browseCmd } from "./browse.ts";
+export { checkoutCmd } from "./checkout.ts";
+export { cloneCmd } from "./clone.ts";
+export { createCmd } from "./create.ts";
+export { pullCmd } from "./pull.ts";
+export { pushCmd } from "./push.ts";
+export { statusCmd } from "./status.ts";
+export { watchCmd } from "./watch.ts";
