--- src/cmd/root/root.ts
@@ -1,24 +1,0 @@
-import { Command } from "@cliffy/command";
-import manifest from "../../../deno.json" with { type: "json" };
-import * as gitCmds from "~/cmd/git/mod.ts";
-import { createCmd } from "~/cmd/root/create.ts";
-import { watchCmd } from "~/cmd/root/watch.ts";
-import { browseCmd } from "~/cmd/root/browse.ts";
-
-const cmd = new Command()
-  .name("vt")
-  .version(manifest.version)
-  .help({ colors: Deno.stdout.isTerminal() })
-  .action(() => cmd.showHelp());
-
-cmd.command("clone", gitCmds.cloneCmd);
-cmd.command("push", gitCmds.pushCmd);
-cmd.command("pull", gitCmds.pullCmd);
-cmd.command("status", gitCmds.statusCmd);
-cmd.command("branch", gitCmds.branchCmd);
-cmd.command("checkout", gitCmds.checkoutCmd);
-cmd.command("watch", watchCmd);
-cmd.command("browse", browseCmd);
-cmd.command("create", createCmd);
-
-export { cmd };
